at work

+++ iOS/G3MiOSSDK/Commons/Basic/ShortBufferElevationData.hpp
@@ -55,7 +55,7 @@ public:
   bool hasNoData() const { return _hasNoData;}
 
   bool hasChildren() const { if (_children > 0) return true; return false; }
-  int getSimilarity() const {return _similarity; }
+  short getSimilarity() const {return _similarity; }
 
 };
